Conflicts:
	CMakeLists.txt
	cmake/Config.cmake
	cmake/Macros.cmake
	src/SFML/Graphics/CircleShape.cpp
	src/SFML/Graphics/ConvexShape.cpp
	src/SFML/Graphics/TextureSaver.cpp
	src/SFML/Graphics/TextureSaver.hpp
	src/SFML/System/Unix/ClockImpl.hpp
	src/SFML/System/Unix/SleepImpl.hpp
	src/SFML/System/Win32/SleepImpl.cpp

+++ src/SFML/Window/GlResource.cpp
@@ -31,13 +31,10 @@
 #include <SFML/System/Lock.hpp>
 
 
-////////////////////////////////////////////////////////////
-// Private data
-////////////////////////////////////////////////////////////
 namespace
 {
     // OpenGL resources counter and its mutex
-    unsigned long count = 0;
+    unsigned int count = 0;
     sf::Mutex mutex;
 }
